feat: add monitoring use only disclaimer modal
