+++ app/src/components/InfoCard/InfoCard.module.scss
@@ -1,6 +1,6 @@
-@import '../../../styles/_variables.scss';
-@import '../../../styles/_mixins.scss';
-@import '../../../styles/_icons.scss';
+@import '../../styles/variables';
+@import '../../styles/mixins';
+@import '../../styles/icons';
 
 .infoCardComponent {
   width: 100%;
