+++ app/src/components/InfoCard/InfoCard.tsx
@@ -1,10 +1,10 @@
-import React, { FC } from 'react';
+import { FC } from 'react';
 import { Tooltip } from 'antd';
 import { InfoCircleFilled } from '@ant-design/icons';
 import c from 'classnames';
 import styles from './InfoCard.module.scss';
 import { getTokenIconClass } from '../../utils/utils';
-import { Button } from '../../components/UI/Button/Button';
+import { Button } from '../UI/Button/Button';
 
 interface InfoCardProps {
   title: string;
